requirements.txt
python-decouple
discord.py
openai